requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
requests